WIP: mend more SQL incompatibilities to get more tests to pass, e.g. auth::auth_test::test_auth_change_email_flow is passing now.

+++ crates/core/migrations/pg_main/U2__file_deletions.sql
@@ -3,16 +3,16 @@
 -- Triggers are being used to populate this deletion log. Keeping a log of
 -- pending deletions also lets us retry deletions in case of transient errors.
 CREATE TABLE _file_deletions (
-  id                           INTEGER PRIMARY KEY NOT NULL,
-  deleted                      INTEGER NOT NULL DEFAULT (UNIXEPOCH()),
+  id                           INT8 PRIMARY KEY NOT NULL,
+  deleted                      INT8 NOT NULL DEFAULT (UNIXEPOCH()),
 
   -- Cleanup metadata
-  attempts                     INTEGER NOT NULL DEFAULT 0,
+  attempts                     INT8 NOT NULL DEFAULT 0,
   errors                       TEXT,
 
   -- Which record contained the file.
   table_name                   TEXT NOT NULL,
-  record_rowid                 INTEGER NOT NULL,
+  record_rowid                 INT8 NOT NULL,
   column_name                  TEXT NOT NULL,
 
   -- File metadata, including id (path).
